chore: Reorganize folders
