+++ packages/components/src/DateTimePickers/DateTimePicker/DateTimePicker.component.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import theme from './DateTimePicker.scss';
-import DateTimeView from '../DateTimeView';
-import MonthYearView from '../MonthYearView';
+import DateTimeView from '../views/DateTimeView';
+import MonthYearView from '../views/MonthYearView';
 
 function DateTimePicker(props) {
 	const isCalendarView = true;
