+++ packages/components/src/DateTimePickers/shared/IconButton/IconButton.component.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import Icon from '../../Icon';
+import Icon from '../../../Icon';
 import theme from './IconButton.scss';
 
 function IconButton(props) {
