+++ packages/components/src/DateTimePickers/subpickers/DatePicker/DatePicker.scss
@@ -1,4 +1,4 @@
-@import '../common';
+@import '../../shared/common';
 
 .container {
     @include main;
