+++ packages/components/src/DateTimePickers/subpickers/DatePicker/DayPickerAction/DayPickerAction.scss
@@ -1,4 +1,4 @@
-@import '../../common';
+@import '../../../shared/common';
 
 .action {
     background: none;
