+++ packages/components/src/DateTimePickers/subpickers/MonthPicker/MonthPicker.scss
@@ -1,5 +1,5 @@
 
-@import '../common';
+@import '../../shared/common';
 
 .container {
     display: flex;
