+++ packages/components/src/DateTimePickers/views/DateTimeView/DateTimeView.scss
@@ -1,4 +1,4 @@
-@import '../common';
+@import '../../shared/common';
 
 .body {
     display: flex;
