+++ packages/components/src/DateTimePickers/views/MonthYearView/MonthYearView.scss
@@ -1,5 +1,5 @@
 
-@import '../common';
+@import '../../shared/common';
 
 .action-left {
     .action-left-icon {
