+++ packages/components/src/DateTimePickers/views/common/HeaderTitle/HeaderTitle.scss
@@ -1,4 +1,4 @@
-@import '../common';
+@import '../../../shared/common';
 
 .common {
     @include main;
