+++ packages/components/src/DateTimePickers/views/common/HeaderTitle/index.js
