+++ packages/components/src/DateTimePickers/views/common/ViewLayout/index.js
